+++ src/components/PloggingList/FilteringBox.js
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import ic_reset from "../../assets/PloggingList/ic_reset.png";
@@ -8,10 +8,25 @@ const FilteringBox = ({
   exceptEnded,
   setExceptEnded,
   selectedBtn,
-  btnList,
+  setSelectedBtn,
+  searchParam, // 필터링 url (?뒤에 나오는 주소)
 }) => {
   const currentUrl = "/plogging-list";
+
   const navigate = useNavigate();
+
+  // 필터링 버튼 리스트
+  const btnList = [
+    { id: 0, title: "연령무관", keyword: "?age=ANY" },
+    { id: 1, title: "성별무관", keyword: "?gender=ANY" },
+    { id: 2, title: "여성만", keyword: "?gender=FEMALE" },
+    { id: 3, title: "남성만", keyword: "?gender=MALE" },
+    {
+      id: 4,
+      title: "반려동물과 함께",
+      keyword: "?withPet=TRUE",
+    },
+  ];
 
   // 필터링 기능
   const toggleButtonClick = (id) => {
@@ -19,6 +34,20 @@ const FilteringBox = ({
     // url에 필터링 요소를 붙이기
     navigate(currentUrl + element.keyword);
   };
+
+  useEffect(() => {
+    // url 검색을 통해 필터링 요소가 있는지 확인
+    const findItem = btnList.find((item) => item.keyword === searchParam);
+
+    if (searchParam === "" || findItem) {
+      // 필터링 요소가 없거나 (전체 list), 올바른 필터링 요소가 있을 경우
+      setSelectedBtn(searchParam === "" ? null : findItem);
+    } else {
+      // 잘못된 키워드로 검색했을 경우
+      alert("올바른 키워드로 검색해주세요.");
+      window.location.href = currentUrl;
+    }
+  }, [searchParam]);
 
   return (
     <>
