+++ src/pages/Plogging/PloggingListPage.js
@@ -11,45 +11,18 @@ import { getPostsList, getPostsListUnAuth } from "../../api/posts";
 
 const PloggingListPage = () => {
   const location = useLocation();
-  const searchParam = location.search;
+  const searchParam = location.search; // 필터링 url (?뒤에 나오는 주소)
 
   const isLogin = !!localStorage.getItem("juptoken"); // 로그인 여부
 
   // 모집마감 제외 여부
   const [exceptEnded, setExceptEnded] = useState(true);
 
-  // 필터링 버튼
-  const btnList = [
-    { id: 0, title: "연령무관", keyword: "?age=ANY" },
-    { id: 1, title: "성별무관", keyword: "?gender=ANY" },
-    { id: 2, title: "여성만", keyword: "?gender=FEMALE" },
-    { id: 3, title: "남성만", keyword: "?gender=MALE" },
-    {
-      id: 4,
-      title: "반려동물과 함께",
-      keyword: "?withPet=TRUE",
-    },
-  ];
-
   // 선택된 버튼
   const [selectedBtn, setSelectedBtn] = useState(null);
 
   // api로 받아온 data
   const [pageData, setPageData] = useState([]);
-
-  useEffect(() => {
-    // url 검색을 통해 필터링 요소가 있는지 확인
-    const findItem = btnList.find((item) => item.keyword === searchParam);
-
-    if (searchParam === "" || findItem) {
-      // 필터링 요소가 없거나 (전체 list), 올바른 필터링 요소가 있을 경우
-      setSelectedBtn(searchParam === "" ? null : findItem);
-    } else {
-      // 잘못된 키워드로 검색했을 경우
-      alert("올바른 키워드로 검색해주세요.");
-      window.location.href = "/plogging-list";
-    }
-  }, [location.search]);
 
   // api 가져오는 함수
   const getData = async () => {
@@ -81,12 +54,13 @@ const PloggingListPage = () => {
   return (
     <>
       <Fixed>
-        <Header title={"플로깅 이벤트"}/>
+        <Header title={"플로깅 이벤트"} />
         <FilteringBox
           exceptEnded={exceptEnded}
           setExceptEnded={setExceptEnded}
           selectedBtn={selectedBtn}
-          btnList={btnList}
+          setSelectedBtn={setSelectedBtn}
+          searchParam={searchParam}
         />
       </Fixed>
